mnt: allow external slaves too
In particular, if the thing in criu's namespace is also a slave of
something in an external mount namespace (as it is in the tests), we won't
see the master, just the external slave. However, this also counts as an
"external master" because it is an external bind mount, just from something
that was itself a slave.

I'm not sure why the test broke, but it works on ubuntu 16.04 and fails on
16.10, so I suppose something with the debugfs mount configuration has
changed. I can look into it more if we're curious, but this patch fixes it
and should probably go in anyway.

Closes #223

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>