net: add support for macvlan link types
On Thu, Oct 13, 2016 at 04:21:21PM -0600, Tycho Andersen wrote:
> On Thu, Oct 13, 2016 at 11:36:03AM -0600, Tycho Andersen wrote:
> > On Thu, Oct 13, 2016 at 11:16:18AM -0600, Tycho Andersen wrote:
> > > Here's an updated patch with a feature check for
> > > linux/net_namespace.h.
> >
> > And here's one that doesn't include headers that clash. Fourth time is
> > a charm, right? :)
>
> Seems all of that worked, and now there's more stuff that's not
> present in older headers. Here's a v5 that will hopefully compile
> everywhere :)

Bah. This time I actually built net.o against the 3.2 kernel headers,
which don't have IFLA_MACVLAN_FLAGS. Hopefully this one will finally
work.

Tycho

>From 23bf23db74e21bca1bafb3ac4afdbbe979bb8947 Mon Sep 17 00:00:00 2001
From: Tycho Andersen <tycho.andersen@canonical.com>
Date: Mon, 13 Jun 2016 23:24:48 +0000
Subject: [PATCH v6 5/6] net: add support for macvlan link types

While this is in principle similar to how veths are handled, we have to do
things in two different ways depending on whether or not there is a user
namespace involved, because there is no way to ask the kernel to attach a
macvlan NIC to a device in a net ns that we don't have CAP_NET_ADMIN in.

So we do it in two ways:

a. If we are in a user namespace, we create the device in usernsd and use
   IFLA_NET_NS_PID to set the netns which it should be created in (saving
   us a "move into this netns" step).

b. If we aren't in a user namespace, we could still be in a net namespace,
   so we use IFLA_LINK_NETNSID to set namespace that the i/o device will be
   in. Then we open a netlink socket from criu's netns and use
   IFLA_NET_NS_FD to tell the kernel to create the macvlan device in the
   target's namespace.

v2: * s/CLONE_NEWNET/CLONE_NEWUSER
    * Don't bother to dump IFLA_LINK and IFLA_LINK_NETNSID. Although we
      need to provide these on restore, there's no kernel interface that
      persists these. To populate IFLA_LINK, we require users pass
      --macvlan-pair, and we create a NETNSID relation as needed and pass
      that in for macvlan links (although this infrastructure could be used
      elsewhere for links that need it in the future, since is in the
      hoisted populate_newlink_req()).
    * use new external command instead of creating a --macvlan-pair option

v3: add a feature check for linux/net_namespace.h, since not every arch in
    travis has this (new-ish) header

v4: * include sys/types.h instead of linux/if.h to get IFF_UP flag
    * remove old doc addition about --macvlan-pair option

v5: define IFLA_LINK_NETNSID and RTM_NEWNSID if they don't exist

v6: define IFLA_MACVLAN_FLAGS and bump the size of IFLA_MACVLAN_MAX when
    necessary

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>