lsm: support checkpoint/restore of stacked apparmor profiles
Support for apparmor namespaces and stacking is coming to Ubuntu kernels in
16.10, and should hopefully be upstreamed Soon (TM) :).

Apparmor namespaces aren't traditional namespaces implemented with CLONE_,
but are implemented via a wholly separate namespace mechanism whose primary
interface is mkdir/rmdir in securityfs. Apparmor keeps the loaded binary
profile bits around (much like seccomp), and makes those available in
securityfs as well for the purposes of checkpoint resotre.

This patch also implements support for "stacked" profiles. Profile stacking
is simply the idea that more than one profile can be applied to the same
task. From CRIU's perspective this mostly looks the same as a single
profile: stacking is simply implemented via some sigils in between profile
names exported from /proc/self/attr/current.

The basic idea for checkpoint restore of stacking and namespaces is similar
to how cgroups are done: we can restore the apparmor namespace and profile
blobs independently of the tasks, and then at the end we can just set the
task's label appropriately. This means the code that moves tasks under a
label stays the same, and the only new code is the stuff that dumps and
restores the policy blobs that are in the namespace that were loaded by the
container.

v2: dump profiles in timestamp order instead of name order; the timestamp
    indicates what order profiles are installed in, and users can
    potentially reload a profile that changes others, so we should insert
    the blobs in timestamp order

v3: better patch notes (hopefully)

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>