lsm: support checkpoint/restore of stacked apparmor profiles
Support for apparmor namespaces and stacking is coming to Ubuntu kernels in
16.10, and should hopefully be upstreamed Soon (TM) :).

The basic idea is similar to how cgroups are done: we can restore the
apparmor namespace and profile blobs independently of the tasks, and then
at the end we can just set the task's label appropriately. This means the
code that moves tasks under a label stays the same, and the only new code
is the stuff that dumps and restores the policy blobs that are in the
namespace that were loaded by the container.

v2: dump profiles in timestamp order instead of name order; the timestamp
    indicates what order profiles are installed in, and users can
    potentially reload a profile that changes others, so we should insert
    the blobs in timestamp order
v3: support profiles with /s in them

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>