lsm: add support for rewriting apparmor profiles
--lsm-profile is used to rewrite some apparmor profiles which might contain
the name or runtime path of the container engine, and thus might be
different on restore. Let's extend its support to apparmor profiles.

Note that if there is nesting beyond what is specified by --lsm-profile.
The assumption here is that it is only necessary to rewrite some prefix of
the full path, and that stuff below that in the tree was set up by whatever
was in the container and should be left alone.

v2: fix incredibly wrong logic for rewriting the actual computed profile
    names :)

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>