sigframe: Move sigframe-related code into sigframe.*
There's a lot of code making sigframes for PIE spread all over the
asm/restorer.h and arch/crtools.c. This patch collects it all into
asm/sigframe.h and arch/sigframe.c for better modularity and for
the sake of compel.

The patch is huge, but it just moves the code around (making u32/u64
types conversions to uint32/64_t-s where appropriate).

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-and-ported-on-dev-by: Pavel Emelyanov <xemul@virtuozzo.com>