mm: Record page size in checkpoint and check it at run time
Page size is a build time kernel configuration option on 64-bit ARM and
PowerPC. Record the page size when a checkpoint is dumped, and check that
it is the same during restore.

With a run-time check in place, remove the build-time check that was
unnecessarily preventing CRIU to be built with PAGE_SHIFT=16 (64K) on arm64 and
PAGE_SHIFT=12 (4K) on ppc64, for example.

Signed-off-by: Christopher Covington <cov@codeaurora.org>
Reviewed-by: Andrei Vagin <avagin@gmail.com>