net: Make criu do not fail on recent iproute2
On 02.11.2016 12:24, Andrey Melnikov wrote:
> 2016-11-02 12:15 GMT+03:00 Kirill Tkhai <ktkhai@virtuozzo.com>:
>> Since iprule commit 67a990b81126 command "ip rule del" is not working anymore:
>>
>>     iproute: disallow ip rule del without parameters
>>
>>     Disallow run `ip rule del` without any parameter to avoid delete any first
>>     rule from table.
>>
>>     Signed-off-by: Andrey Jr. Melnikov <temnota.am@gmail.com>
>>
>> So, criu restore fails with:
>>
>>     Error (criu/net.c:1277): IP tool failed on rule delete
>>
>> Fix that by explicit passing of rule's table.
>>
>> v2: Use "ip rule flush" to main's and default's table rules at once.
>
> Also, another stupid question - why not invoke `ip -batch -` with
> popen() and push commands to pipe?

We have many other places where ip is called in explicit way for a single action.
It may have a sense rework them all using a single pipe. Pavel, how do you think?

Anyway, this patch is a fix to make criu working on recent iproute2. And it's not
for reworking. I'm not sure we have enough time to do this now. But... Andrew,
do you want to participate in open source project? :) If Pavel thinks we should
rework all ip calls using a single pipe.

Using a pipe for this place only is below. Chose a version you like more...

[PATCH]net: Make criu do not fail on recent iproute2

    Since iprule commit 67a990b81126 command "ip rule del" is not working anymore:

        iproute: disallow ip rule del without parameters

        Disallow run `ip rule del` without any parameter to avoid delete any first
        rule from table.

        Signed-off-by: Andrey Jr. Melnikov <temnota.am@gmail.com>

    So, criu restore fails with:

        Error (criu/net.c:1277): IP tool failed on rule delete

    Fix that by explicit passing of rule's table.

    v2: Use "ip rule flush" to main's and default's table rules at once.
    v3: Use "ip -batch"

    Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>