infect: Introduce infect_ctx
This structure will become the part of compel A[PB]I and
will be used by compel clients as configuration structure.

For now put there the pointer on a socket living in victim's
netns. If empty (-1), compel will create this socket its own,
but CRIU already creates such in an effective manner.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>