lazy-pages: fix searching for the page at #PF time
After commit a97d6d3f1609 (pagemap: replace seek_page with seek_pagemap
method), uffd only searches the pagemap containing the faulting page, but
it not for the page itself. For local restore it causes wrong data to be
read from pages*img and subsequent crash of the restored process.
Adding a call to pr->skip_pages fixes the problem.

Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>