mnt: try to split a mount tree to restore over-mounted mounts
If a mount overmounts something else, we can try to restore it
separetly and then move it to the right places after restoring
all mounts.

In this patch if we see that a mount is overmounts something,
we create a new directory in the root yard and restore this
mount and its sub-tree in this directory.

https://bugs.openvz.org/browse/OVZ-6778

v2: add more comments
    rename roots_mp into roots_yard_mp
v3: rename functions

Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>