Makefile.install: DESTDIR should be empty
DESTDIR is mostly used by packagers (such as from rpm spec
or deb rules), or just to try installation, and normally
should be left unset/empty.

Setting it to / looks weird to me. Besides, by leaving it empty
we are making sure everything works if DESTDIR is not set.

Fix INSTALL.md accordingly.

CC: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>