test: Fix sockets03 default SIGPIPE handler not being set to IGN
By default, sockets03 doesn't set SIGPIPE handler to SIG_IGN. When run
by zdtm.py, it doesn't cause problem, because it inherits signal
handler. However, when run via vzt-cpt (which uses system() to launch
tests), it receives default handler (which would terminate test). As
result, test is guaranteed to fail, because it will attempt to write to
shutdown receiver and get SIGPIPE in the face.

PSBM-55941

Signed-off-by: Vitaly Ostrosablin <vostrosablin@virtuozzo.com>