page-server: Allow blocking on pipe
This splice tries to get pages from socket into local pipe to
splice them into images later. The data on the socket may not
be there by the time we get to this splice, so there's no reason
to force non-blocking IO here.

travis-ci: success for page-server: Allow blocking on pipe
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>