Revert "page-server: Allow blocking on pipe"
This reverts commit 00e25daaf2bddf8ff998b7ef0e0ef70b988cd660.

Actually we can't block on pipe, because the page-server has
only one thread. It splices data from socket to pipe and then
reads data from this pipe. We are tring to calculete a chunk size
to not block on pipe, but this chunk size isn't accurate if
spliced data are not allign to a page size.

https://github.com/xemul/criu/issues/265
Signed-off-by: Andrei Vagin <avagin@virtuozzo.com>