dump: add root_only option
This option allows to dump only root of a specified process
tree. This behaviour is useful for dumping master processes,
that fork it's children to do the work. I.e. one of such cases
is php process that forks workers. When one uses this option,
he agrees to possible consequences and is ready to handle
recovering from such restore on a dumpee side.

Signed-off-by: Ruslan Kuprieiev <rkuprieiev@cloudlinux.com>