criu/Makefile: fix criu/arch/* rebuild
Problem: if we do

 touch criu/arch/x86/cpu.c

nothing is being rebuilt.

For detailed analisys, see the previous commit
("criu/Makefile: fix rebuilding criu/pie/pie.lib.a").

Note there are other targets in criu/Makefile where one
might want to add .FORCE to, but it's not needed as they
all depend on another target which itself is marked as .PHONY.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>