lazy-pages: track outstanding page faults
Multithreaded applications may concurrently generate page faults at the
same address, which will cause creation of multiple requests for remote
page, and, consequently, confuse the page server on the dump side.
We can keep track on page fault requests in flight and ensure this way that
we request a page from the remote side only once.

Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>